# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Naturascan API",
        "description": "Naturascan API",
        "contact": {
            "name": "Gilles",
            "email": "gillesakakpo01@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/gps_tracks": {
            "get": {
                "tags": [
                    "GpsTrack"
                ],
                "summary": "List all gps_track species",
                "operationId": "25566d3e97b4e1bf281b0cdadcac3968",
                "responses": {
                    "200": {
                        "description": "List all gps_track species"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "GpsTrack"
                ],
                "summary": "Create a new gps_track species",
                "operationId": "1a3c9f08823dc4b6e139fc108762347a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GpsTrack"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Create a new gps_track species"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/gps_tracks/{id}": {
            "get": {
                "tags": [
                    "GpsTrack"
                ],
                "summary": "Show gps_track species by ID",
                "operationId": "9ac29ebdec3e63ef5c5ef10377b476a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the gps_track species",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show gps_track species by ID"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "GpsTrack"
                ],
                "summary": "Update gps_track species by ID",
                "operationId": "e5a14d26a033ca6b9a20d7233250eed7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the gps_track species",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GpsTrack"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update gps_track species by ID"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "GpsTrack"
                ],
                "summary": "Delete gps_track species by ID",
                "operationId": "5e1485ac0e370e9d412e0f5193e7ad95",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the gps_track species",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete gps_track species by ID"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/observations": {
            "get": {
                "tags": [
                    "Observation"
                ],
                "summary": "List all observation species",
                "operationId": "23ca8dd7b03edddffba2e2ff578ca9b2",
                "responses": {
                    "200": {
                        "description": "List all observation species"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Observation"
                ],
                "summary": "Create a new observation species",
                "operationId": "b3923335c07e2fdf3f812e069a90467c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Observation"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Create a new observation species"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/observations/{id}": {
            "get": {
                "tags": [
                    "Observation"
                ],
                "summary": "Show observation species by ID",
                "operationId": "943b1b4d6a7022840a4aec1e8f74ebf2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the observation species",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show observation species by ID"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Observation"
                ],
                "summary": "Update observation species by ID",
                "operationId": "5ff20536907b3e56378285bfa19166c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the observation species",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Observation"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update observation species by ID"
                    },
                    "404": {
                        "description": "Observation species not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Observation"
                ],
                "summary": "Delete observation species by ID",
                "operationId": "78c5e97deb347d7d8a7781d8d285fa85",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the observation species",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Delete observation species by ID"
                    },
                    "404": {
                        "description": "Observation species not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "GpsTrack": {
                "title": "GpsTrack",
                "description": "GpsTrack",
                "properties": {
                    "id": {
                        "title": "id",
                        "type": "integer"
                    },
                    "longitude": {
                        "title": "longitude",
                        "type": "string"
                    },
                    "shipping_id": {
                        "title": "shipping_id",
                        "type": "integer"
                    },
                    "latitude": {
                        "title": "latitude",
                        "type": "string"
                    },
                    "timestamp": {
                        "title": "timestamp",
                        "type": "string"
                    },
                    "device": {
                        "title": "device",
                        "type": "string"
                    },
                    "inObservation": {
                        "title": "inObservation",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "Observation": {
                "title": "Observation",
                "description": "Observation",
                "properties": {
                    "id": {
                        "title": "id",
                        "type": "integer"
                    },
                    "type": {
                        "title": "type",
                        "type": "string"
                    },
                    "sortie_id": {
                        "title": "sortie_id",
                        "type": "integer"
                    },
                    "animal_id": {
                        "title": "animal_id",
                        "type": "integer"
                    },
                    "bird_id": {
                        "title": "bird_id",
                        "type": "integer"
                    },
                    "waste_id": {
                        "title": "waste_id",
                        "type": "integer"
                    },
                    "created_at": {
                        "title": "created_at",
                        "type": "string"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}